Added test_examples_of_prod_image_building in python (#21097)

+++ airflow/migrations/versions/cf5dc11e79ad_drop_user_and_chart.py
@@ -18,7 +18,7 @@
 """drop_user_and_chart
 
 Revision ID: cf5dc11e79ad
-Revises: 41f5f12752f8
+Revises: 03afc6b6f902
 Create Date: 2019-01-24 15:30:35.834740
 
 """
@@ -29,7 +29,7 @@ from sqlalchemy.engine.reflection import Inspector
 
 # revision identifiers, used by Alembic.
 revision = 'cf5dc11e79ad'
-down_revision = '41f5f12752f8'
+down_revision = '03afc6b6f902'
 branch_labels = None
 depends_on = None
 
@@ -37,18 +37,27 @@ depends_on = None
 def upgrade():
     # We previously had a KnownEvent's table, but we deleted the table without
     # a down migration to remove it (so we didn't delete anyone's data if they
-    # were happing to use the feature.
+    # were happening to use the feature.
     #
     # But before we can delete the users table we need to drop the FK
 
     conn = op.get_bind()
     inspector = Inspector.from_engine(conn)
+    tables = inspector.get_table_names()
 
-    if 'known_event' in inspector.get_table_names() != 'sqlite':
-        op.drop_constraint('known_event_user_id_fkey', 'known_event')
+    if 'known_event' in tables:
+        for fkey in inspector.get_foreign_keys(table_name="known_event", referred_table="users"):
+            if fkey['name']:
+                with op.batch_alter_table(table_name='known_event') as bop:
+                    bop.drop_constraint(fkey['name'], type_="foreignkey")
 
-    op.drop_table("chart")
-    op.drop_table("users")
+    if "chart" in tables:
+        op.drop_table(
+            "chart",
+        )
+
+    if "users" in tables:
+        op.drop_table("users")
 
 
 def downgrade():
@@ -62,7 +71,7 @@ def downgrade():
         sa.Column('password', sa.String(255)),
         sa.Column('superuser', sa.Boolean(), default=False),
         sa.PrimaryKeyConstraint('id'),
-        sa.UniqueConstraint('username')
+        sa.UniqueConstraint('username'),
     )
 
     op.create_table(
@@ -82,8 +91,11 @@ def downgrade():
         sa.Column('x_is_date', sa.Boolean(), nullable=True),
         sa.Column('iteration_no', sa.Integer(), nullable=True),
         sa.Column('last_modified', sa.DateTime(), nullable=True),
-        sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-        sa.PrimaryKeyConstraint('id')
+        sa.ForeignKeyConstraint(
+            ['user_id'],
+            ['users.id'],
+        ),
+        sa.PrimaryKeyConstraint('id'),
     )
 
     if conn.dialect.name == 'mysql':
